--- khal_config.cfg
@@ -1,14 +1,0 @@
-# app secret key
-SECRET_KEY = '6AA0p[$f//LOLlpw'
-
-# configuration
-
-
-CSRF_ENABLED = True
-DATABASE = "/home/maria/Desktop/ciacicode/khaleesicode/microblog.db"
-USERNAME = "ciacicode"
-PASSWORD = "KLC!blogpass"
-PER_PAGE = 10
-CSS_FRAMEWORK = "bootstrap3"
-LINK_SIZE = "sm"
-SHOW_SINGLE_PAGE = False
